Quote packet-derived names before handing them to the shell
> The following was drafted by Claude (Opus 5), an LLM made by
> Anthropic, at Andy's direction.

compressAddFile passes its last argument to the shell unquoted, because
the reply drivers use it to pass globs. The variable part of those
templates is replyInnerName, built from the packet's own header, and
read.cc passed a bare file name there, so a packet could name itself
"A;cmd" and run a command when its replies were archived. Quote the
names and leave the globs alone.

The interfac.cc cast silences a CodeQL warning about int multiplication
widening to size_t; the product is bounded by the selector count times
the longest label, so it cannot overflow.

+++ interfac/interfac.cc
@@ -192,7 +192,7 @@ int Interface::WarningWindow(const char *warning, const char **selectors,
     // Display each item:
 
     for (p = selectors, curitem = 0; curitem < items; curitem++) {
-        x = curitem * itemlen + ((itemlen - strlen(*p) + 5) >> 1);
+        x = curitem * itemlen + ((itemlen - (int) strlen(*p) + 5) >> 1);
 
         warn.attrib(C_WTEXT);
         warn.put(4, x + 1, *p + 1);
